I have found a low-level read routine overridden by the original Falcon 4.0 developers. Probably at their time this low-level read routine was not included in the VS's library, but now it is, and the old one is not working, thus causing Issue #1, I comment it out, it seems to be working. I also commented out a line in "TexBank.cpp" which sets the "maxCompressedSize" to 0xffffffff causing the following new comand to fail. The application still having problems at this point, the problem seem to be located in "DrawParticleSys.cpp" the function "bool DrawableParticleSys::LoadParameters(void)" I am not quite familar with DirectX stuff, so I am going to leave it to you guys at the moment.
I have found a low-level read routine overridden by the original Falcon 4.0 developers.
Probably at their time this low-level read routine was not included in the VS's library,
but now it is, and the old one is not working, thus causing Issue #1, I comment it out,
it seems to be working.
I also commented out a line in "TexBank.cpp" which sets the "maxCompressedSize" to 0xffffffff
causing the following new comand to fail.
The application still having problems at this point, the problem seem to be located in
"DrawParticleSys.cpp" the function "bool DrawableParticleSys::LoadParameters(void)"
when the while loop goes into "On("size"){}" part.
I am not quite familar with DirectX stuff, so I am going to leave it to you guys at the
moment.

Plutonium

Signed-off-by: Plutonium <zhitaoli.hust@gmail.com>